refactor: split financial advisor into components
